# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Upcreate Creator Intelligence API",
    "description": "Backend API for creator metrics, commerce scoring, brands, and campaigns (MVP).",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "/api/v1", "description": "Relative (same host)" },
    { "url": "http://localhost:4000/api/v1", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/brands": {
      "get": {
        "tags": ["Brands"],
        "summary": "List brands",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BrandListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Brands"],
        "summary": "Create brand",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBrand" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Brand" }
              }
            }
          },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/brands/{id}": {
      "get": {
        "tags": ["Brands"],
        "summary": "Get brand with campaigns",
        "parameters": [{ "$ref": "#/components/parameters/BrandId" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BrandWithCampaigns" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Brands"],
        "summary": "Update brand",
        "parameters": [{ "$ref": "#/components/parameters/BrandId" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateBrand" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Brand" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Brands"],
        "summary": "Delete brand",
        "parameters": [{ "$ref": "#/components/parameters/BrandId" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register (BRAND creates brand row; CREATOR creates creator row; ADMIN needs adminRegisterSecret)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User + JWT" },
          "403": { "description": "Admin registration blocked" },
          "409": { "description": "Email taken" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "User + JWT" }, "401": { "description": "Invalid credentials" } }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current user from JWT",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "User profile" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/admin/audit-logs": {
      "get": {
        "tags": ["Admin"],
        "summary": "List audit logs (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "actorUserId", "in": "query", "schema": { "type": "string" } },
          { "name": "entityType", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": { "200": { "description": "Paginated audit logs" }, "403": { "description": "Forbidden" } }
      }
    },
    "/admin/workflows": {
      "get": {
        "tags": ["Admin"],
        "summary": "List workflow definitions",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "scope", "in": "query", "schema": { "type": "string", "enum": ["CAMPAIGN","CAMPAIGN_CREATOR","BRAND","CREATOR"] } }],
        "responses": { "200": { "description": "Workflow definitions" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create new workflow version",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["scope", "definition"],
                "properties": {
                  "scope": { "type": "string", "enum": ["CAMPAIGN","CAMPAIGN_CREATOR","BRAND","CREATOR"] },
                  "definition": { "type": "object" },
                  "setActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Invalid definition" } }
      }
    },
    "/admin/workflows/{id}/activate": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Activate a workflow version",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Activated" } }
      }
    },
    "/admin/workflows/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a workflow definition",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/admin/app-users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List app users",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Paginated users" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create app user (with optional brand/creator)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "role": { "type": "string", "enum": ["ADMIN","CREATOR","BRAND"] },
                  "brandName": { "type": "string" },
                  "creatorName": { "type": "string" },
                  "brandExtras": { "type": "object" },
                  "creatorExtras": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "409": { "description": "Email already exists" } }
      }
    },
    "/admin/app-users/{id}/password": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Reset user password",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["password"], "properties": { "password": { "type": "string", "minLength": 8 } } } } }
        },
        "responses": { "200": { "description": "Password updated" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login via POST /auth/login; send Authorization: Bearer <token>"
      }
    },
    "parameters": {
...
      "BrandId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "schemas": {
...
      "UserRole": { "type": "string", "enum": ["BRAND", "CREATOR", "ADMIN"] },
...
      "Brand": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "website": { "type": "string", "nullable": true },
          "category": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email" },
          "notes": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "BrandWithCampaigns": {
        "allOf": [
          { "$ref": "#/components/schemas/Brand" },
          {
            "type": "object",
            "properties": {
              "campaigns": { "type": "array", "items": { "type": "object" } }
            }
          }
        ]
      },
      "BrandListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Brand" } },
          "meta": { "$ref": "#/components/schemas/PaginationMeta" }
        }
      },
      "CreateBrand": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "email": { "type": "string", "format": "email" },
          "website": { "type": "string", "format": "uri" },
          "category": { "type": "string" },
          "notes": { "type": "string" }
        }
      },
      "UpdateBrand": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "website": { "type": "string", "format": "uri" },
          "category": { "type": "string" },
          "notes": { "type": "string" }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "role"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "brandName": { "type": "string" },
          "creatorName": { "type": "string" },
          "adminRegisterSecret": { "type": "string", "description": "Required for ADMIN if ADMIN_REGISTER_SECRET is set" }
        }
      }
    }
  }
}